+++ zenoh/src/net/routing/dispatcher/face.rs
@@ -22,10 +22,8 @@ use tokio_util::sync::CancellationToken;
 use zenoh_protocol::{
     core::{ExprId, WhatAmI, ZenohId},
     network::{
-        declare::ext,
         interest::{InterestId, InterestMode, InterestOptions},
-        Declare, DeclareBody, DeclareFinal, Mapping, Push, Request, RequestId, Response,
-        ResponseFinal,
+        Mapping, Push, Request, RequestId, Response, ResponseFinal,
     },
     zenoh::RequestBody,
 };
@@ -35,15 +33,17 @@ use zenoh_transport::multicast::TransportMulticast;
 #[cfg(feature = "stats")]
 use zenoh_transport::stats::TransportStats;
 
-use super::{super::router::*, resource::*, tables, tables::TablesLock};
+use super::{
+    super::router::*,
+    interests::{declare_interest, undeclare_interest},
+    resource::*,
+    tables::{self, TablesLock},
+};
 use crate::{
     api::key_expr::KeyExpr,
     net::{
         primitives::{McastMux, Mux, Primitives},
-        routing::{
-            interceptor::{InterceptorTrait, InterceptorsChain},
-            RoutingContext,
-        },
+        routing::interceptor::{InterceptorTrait, InterceptorsChain},
     },
 };
 
@@ -212,57 +212,17 @@ impl Primitives for Face {
     fn send_interest(&self, msg: zenoh_protocol::network::Interest) {
         let ctrl_lock = zlock!(self.tables.ctrl_lock);
         if msg.mode != InterestMode::Final {
-            if msg.options.keyexprs() && msg.mode != InterestMode::Current {
-                register_expr_interest(
-                    &self.tables,
-                    &mut self.state.clone(),
-                    msg.id,
-                    msg.wire_expr.as_ref(),
-                );
-            }
-            if msg.options.subscribers() {
-                declare_sub_interest(
-                    ctrl_lock.as_ref(),
-                    &self.tables,
-                    &mut self.state.clone(),
-                    msg.id,
-                    msg.wire_expr.as_ref(),
-                    msg.mode,
-                    msg.options.aggregate(),
-                );
-            }
-            if msg.options.queryables() {
-                declare_qabl_interest(
-                    ctrl_lock.as_ref(),
-                    &self.tables,
-                    &mut self.state.clone(),
-                    msg.id,
-                    msg.wire_expr.as_ref(),
-                    msg.mode,
-                    msg.options.aggregate(),
-                );
-            }
-            if msg.mode != InterestMode::Future {
-                self.state.primitives.send_declare(RoutingContext::new_out(
-                    Declare {
-                        interest_id: Some(msg.id),
-                        ext_qos: ext::QoSType::DECLARE,
-                        ext_tstamp: None,
-                        ext_nodeid: ext::NodeIdType::DEFAULT,
-                        body: DeclareBody::DeclareFinal(DeclareFinal),
-                    },
-                    self.clone(),
-                ));
-            }
-        } else {
-            unregister_expr_interest(&self.tables, &mut self.state.clone(), msg.id);
-            undeclare_sub_interest(
+            declare_interest(
                 ctrl_lock.as_ref(),
                 &self.tables,
                 &mut self.state.clone(),
                 msg.id,
+                msg.wire_expr.as_ref(),
+                msg.mode,
+                msg.options,
             );
-            undeclare_qabl_interest(
+        } else {
+            undeclare_interest(
                 ctrl_lock.as_ref(),
                 &self.tables,
                 &mut self.state.clone(),
