+++ zenoh/src/net/routing/dispatcher/queries.rs
@@ -25,7 +25,6 @@ use zenoh_protocol::{
     core::{key_expr::keyexpr, Encoding, WireExpr},
     network::{
         declare::{ext, queryable::ext::QueryableInfoType, QueryableId},
-        interest::{InterestId, InterestMode},
         request::{
             ext::{BudgetType, TargetType, TimeoutType},
             Request, RequestId,
@@ -43,87 +42,6 @@ use super::{
     tables::{NodeId, RoutingExpr, Tables, TablesLock},
 };
 use crate::net::routing::{hat::HatTrait, RoutingContext};
-
-#[allow(clippy::too_many_arguments)] // TODO refactor
-pub(crate) fn declare_qabl_interest(
-    hat_code: &(dyn HatTrait + Send + Sync),
-    tables: &TablesLock,
-    face: &mut Arc<FaceState>,
-    id: InterestId,
-    expr: Option<&WireExpr>,
-    mode: InterestMode,
-    aggregate: bool,
-) {
-    if let Some(expr) = expr {
-        let rtables = zread!(tables.tables);
-        match rtables
-            .get_mapping(face, &expr.scope, expr.mapping)
-            .cloned()
-        {
-            Some(mut prefix) => {
-                tracing::debug!(
-                    "{} Declare qabl interest {} ({}{})",
-                    face,
-                    id,
-                    prefix.expr(),
-                    expr.suffix
-                );
-                let res = Resource::get_resource(&prefix, &expr.suffix);
-                let (mut res, mut wtables) = if res
-                    .as_ref()
-                    .map(|r| r.context.is_some())
-                    .unwrap_or(false)
-                {
-                    drop(rtables);
-                    let wtables = zwrite!(tables.tables);
-                    (res.unwrap(), wtables)
-                } else {
-                    let mut fullexpr = prefix.expr();
-                    fullexpr.push_str(expr.suffix.as_ref());
-                    let mut matches = keyexpr::new(fullexpr.as_str())
-                        .map(|ke| Resource::get_matches(&rtables, ke))
-                        .unwrap_or_default();
-                    drop(rtables);
-                    let mut wtables = zwrite!(tables.tables);
-                    let mut res =
-                        Resource::make_resource(&mut wtables, &mut prefix, expr.suffix.as_ref());
-                    matches.push(Arc::downgrade(&res));
-                    Resource::match_resource(&wtables, &mut res, matches);
-                    (res, wtables)
-                };
-
-                hat_code.declare_qabl_interest(
-                    &mut wtables,
-                    face,
-                    id,
-                    Some(&mut res),
-                    mode,
-                    aggregate,
-                );
-            }
-            None => tracing::error!(
-                "{} Declare qabl interest {} for unknown scope {}!",
-                face,
-                id,
-                expr.scope
-            ),
-        }
-    } else {
-        let mut wtables = zwrite!(tables.tables);
-        hat_code.declare_qabl_interest(&mut wtables, face, id, None, mode, aggregate);
-    }
-}
-
-pub(crate) fn undeclare_qabl_interest(
-    hat_code: &(dyn HatTrait + Send + Sync),
-    tables: &TablesLock,
-    face: &mut Arc<FaceState>,
-    id: InterestId,
-) {
-    tracing::debug!("{} Undeclare qabl interest {}", face, id,);
-    let mut wtables = zwrite!(tables.tables);
-    hat_code.undeclare_qabl_interest(&mut wtables, face, id);
-}
 
 pub(crate) struct Query {
     src_face: Arc<FaceState>,
