+++ zenoh/src/net/routing/hat/client/mod.rs
@@ -26,7 +26,7 @@ use std::{
 use zenoh_config::WhatAmI;
 use zenoh_protocol::network::{
     declare::{queryable::ext::QueryableInfoType, QueryableId, SubscriberId},
-    interest::InterestId,
+    interest::{InterestId, InterestOptions},
     Oam,
 };
 use zenoh_result::ZResult;
@@ -34,6 +34,7 @@ use zenoh_sync::get_mut_unchecked;
 use zenoh_transport::unicast::TransportUnicast;
 
 use self::{
+    interests::interests_new_face,
     pubsub::{pubsub_new_face, undeclare_client_subscription},
     queries::{queries_new_face, undeclare_client_queryable},
 };
@@ -52,6 +53,7 @@ use crate::net::{
     runtime::Runtime,
 };
 
+mod interests;
 mod pubsub;
 mod queries;
 
@@ -100,6 +102,7 @@ impl HatBaseTrait for HatCode {
         _tables_ref: &Arc<TablesLock>,
         face: &mut Face,
     ) -> ZResult<()> {
+        interests_new_face(tables, &mut face.state);
         pubsub_new_face(tables, &mut face.state);
         queries_new_face(tables, &mut face.state);
         Ok(())
@@ -112,6 +115,7 @@ impl HatBaseTrait for HatCode {
         face: &mut Face,
         _transport: &TransportUnicast,
     ) -> ZResult<()> {
+        interests_new_face(tables, &mut face.state);
         pubsub_new_face(tables, &mut face.state);
         queries_new_face(tables, &mut face.state);
         Ok(())
@@ -121,7 +125,7 @@ impl HatBaseTrait for HatCode {
         let mut wtables = zwrite!(tables.tables);
         let mut face_clone = face.clone();
 
-        face_hat_mut!(face).remote_sub_interests.clear();
+        face_hat_mut!(face).remote_interests.clear();
         face_hat_mut!(face).local_subs.clear();
         face_hat_mut!(face).local_qabls.clear();
 
@@ -287,7 +291,7 @@ impl HatContext {
 
 struct HatFace {
     next_id: AtomicU32, // @TODO: manage rollover and uniqueness
-    remote_sub_interests: HashMap<InterestId, Option<Arc<Resource>>>,
+    remote_interests: HashMap<InterestId, (Option<Arc<Resource>>, InterestOptions)>,
     local_subs: HashMap<Arc<Resource>, SubscriberId>,
     remote_subs: HashMap<SubscriberId, Arc<Resource>>,
     local_qabls: HashMap<Arc<Resource>, (QueryableId, QueryableInfoType)>,
@@ -298,7 +302,7 @@ impl HatFace {
     fn new() -> Self {
         Self {
             next_id: AtomicU32::new(0),
-            remote_sub_interests: HashMap::new(),
+            remote_interests: HashMap::new(),
             local_subs: HashMap::new(),
             remote_subs: HashMap::new(),
             local_qabls: HashMap::new(),
