+++ zenoh/src/net/routing/hat/p2p_peer/mod.rs
@@ -43,6 +43,7 @@ use zenoh_transport::unicast::TransportUnicast;
 
 use self::{
     gossip::Network,
+    interests::interests_new_face,
     pubsub::{pubsub_new_face, undeclare_client_subscription},
     queries::{queries_new_face, undeclare_client_queryable},
 };
@@ -65,6 +66,7 @@ use crate::net::{
 };
 
 mod gossip;
+mod interests;
 mod pubsub;
 mod queries;
 
@@ -145,6 +147,7 @@ impl HatBaseTrait for HatCode {
         _tables_ref: &Arc<TablesLock>,
         face: &mut Face,
     ) -> ZResult<()> {
+        interests_new_face(tables, &mut face.state);
         pubsub_new_face(tables, &mut face.state);
         queries_new_face(tables, &mut face.state);
         Ok(())
@@ -173,6 +176,7 @@ impl HatBaseTrait for HatCode {
             );
         }
 
+        interests_new_face(tables, &mut face.state);
         pubsub_new_face(tables, &mut face.state);
         queries_new_face(tables, &mut face.state);
 
@@ -194,9 +198,8 @@ impl HatBaseTrait for HatCode {
         let mut wtables = zwrite!(tables.tables);
         let mut face_clone = face.clone();
 
-        face_hat_mut!(face).remote_sub_interests.clear();
+        face_hat_mut!(face).remote_interests.clear();
         face_hat_mut!(face).local_subs.clear();
-        face_hat_mut!(face).remote_qabl_interests.clear();
         face_hat_mut!(face).local_qabls.clear();
 
         let face = get_mut_unchecked(face);
@@ -388,10 +391,9 @@ impl HatContext {
 
 struct HatFace {
     next_id: AtomicU32, // @TODO: manage rollover and uniqueness
-    remote_sub_interests: HashMap<InterestId, (Option<Arc<Resource>>, bool)>,
+    remote_interests: HashMap<InterestId, (Option<Arc<Resource>>, InterestOptions)>,
     local_subs: HashMap<Arc<Resource>, SubscriberId>,
     remote_subs: HashMap<SubscriberId, Arc<Resource>>,
-    remote_qabl_interests: HashMap<InterestId, Option<Arc<Resource>>>,
     local_qabls: HashMap<Arc<Resource>, (QueryableId, QueryableInfoType)>,
     remote_qabls: HashMap<QueryableId, Arc<Resource>>,
 }
@@ -400,10 +402,9 @@ impl HatFace {
     fn new() -> Self {
         Self {
             next_id: AtomicU32::new(0),
-            remote_sub_interests: HashMap::new(),
+            remote_interests: HashMap::new(),
             local_subs: HashMap::new(),
             remote_subs: HashMap::new(),
-            remote_qabl_interests: HashMap::new(),
             local_qabls: HashMap::new(),
             remote_qabls: HashMap::new(),
         }
