+++ zenoh/src/net/routing/hat/p2p_peer/pubsub.rs
@@ -25,7 +25,6 @@ use zenoh_protocol::{
             DeclareSubscriber, SubscriberId, UndeclareSubscriber,
         },
         interest::{InterestId, InterestMode, InterestOptions},
-        Interest,
     },
 };
 use zenoh_sync::get_mut_unchecked;
@@ -35,7 +34,7 @@ use crate::{
     key_expr::KeyExpr,
     net::routing::{
         dispatcher::{
-            face::{FaceState, InterestState},
+            face::FaceState,
             resource::{NodeId, Resource, SessionContext},
             tables::{Route, RoutingExpr, Tables},
         },
@@ -78,14 +77,16 @@ fn propagate_simple_subscription_to(
             ));
         } else {
             let matching_interests = face_hat!(dst_face)
-                .remote_sub_interests
+                .remote_interests
                 .values()
-                .filter(|si| si.0.as_ref().map(|si| si.matches(res)).unwrap_or(true))
+                .filter(|(r, o)| {
+                    o.subscribers() && r.as_ref().map(|r| r.matches(res)).unwrap_or(true)
+                })
                 .cloned()
-                .collect::<Vec<(Option<Arc<Resource>>, bool)>>();
+                .collect::<Vec<(Option<Arc<Resource>>, InterestOptions)>>();
 
-            for (int_res, aggregate) in matching_interests {
-                let res = if aggregate {
+            for (int_res, options) in matching_interests {
+                let res = if options.aggregate() {
                     int_res.as_ref().unwrap_or(res)
                 } else {
                     res
@@ -342,7 +343,7 @@ pub(super) fn pubsub_new_face(tables: &mut Tables, face: &mut Arc<FaceState>) {
         let sub_info = SubscriberInfo {
             reliability: Reliability::Reliable, // @TODO compute proper reliability to propagate from reliability of known subscribers
         };
-        for mut src_face in tables
+        for src_face in tables
             .faces
             .values()
             .cloned()
@@ -357,33 +358,6 @@ pub(super) fn pubsub_new_face(tables: &mut Tables, face: &mut Arc<FaceState>) {
                     &mut src_face.clone(),
                 );
             }
-            if face.whatami == WhatAmI::Router {
-                for (res, _) in face_hat_mut!(&mut src_face).remote_sub_interests.values() {
-                    let id = face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
-                    let options = InterestOptions::KEYEXPRS + InterestOptions::SUBSCRIBERS;
-                    get_mut_unchecked(face).local_interests.insert(
-                        id,
-                        InterestState {
-                            options,
-                            res: res.as_ref().map(|res| (*res).clone()),
-                            finalized: false,
-                        },
-                    );
-                    let wire_expr = res.as_ref().map(|res| Resource::decl_key(res, face));
-                    face.primitives.send_interest(RoutingContext::with_expr(
-                        Interest {
-                            id,
-                            mode: InterestMode::CurrentFuture,
-                            options,
-                            wire_expr,
-                            ext_qos: ext::QoSType::DECLARE,
-                            ext_tstamp: None,
-                            ext_nodeid: ext::NodeIdType::DEFAULT,
-                        },
-                        res.as_ref().map(|res| res.expr()).unwrap_or_default(),
-                    ));
-                }
-            }
         }
     }
     // recompute routes
@@ -392,38 +366,105 @@ pub(super) fn pubsub_new_face(tables: &mut Tables, face: &mut Arc<FaceState>) {
     update_data_routes_from(tables, &mut tables.root_res.clone());
 }
 
-impl HatPubSubTrait for HatCode {
-    fn declare_sub_interest(
-        &self,
-        tables: &mut Tables,
-        face: &mut Arc<FaceState>,
-        id: InterestId,
-        res: Option<&mut Arc<Resource>>,
-        mode: InterestMode,
-        aggregate: bool,
-    ) {
-        if mode.current() && face.whatami == WhatAmI::Client {
-            let interest_id = (!mode.future()).then_some(id);
-            let sub_info = SubscriberInfo {
-                reliability: Reliability::Reliable, // @TODO compute proper reliability to propagate from reliability of known subscribers
-            };
-            if let Some(res) = res.as_ref() {
-                if aggregate {
-                    if tables.faces.values().any(|src_face| {
-                        src_face.id != face.id
-                            && face_hat!(src_face)
-                                .remote_subs
-                                .values()
-                                .any(|sub| sub.context.is_some() && sub.matches(res))
-                    }) {
+pub(super) fn declare_sub_interest(
+    tables: &mut Tables,
+    face: &mut Arc<FaceState>,
+    id: InterestId,
+    res: Option<&mut Arc<Resource>>,
+    mode: InterestMode,
+    aggregate: bool,
+) {
+    if mode.current() && face.whatami == WhatAmI::Client {
+        let interest_id = (!mode.future()).then_some(id);
+        let sub_info = SubscriberInfo {
+            reliability: Reliability::Reliable, // @TODO compute proper reliability to propagate from reliability of known subscribers
+        };
+        if let Some(res) = res.as_ref() {
+            if aggregate {
+                if tables.faces.values().any(|src_face| {
+                    src_face.id != face.id
+                        && face_hat!(src_face)
+                            .remote_subs
+                            .values()
+                            .any(|sub| sub.context.is_some() && sub.matches(res))
+                }) {
+                    let id = if mode.future() {
+                        let id = face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
+                        face_hat_mut!(face).local_subs.insert((*res).clone(), id);
+                        id
+                    } else {
+                        0
+                    };
+                    let wire_expr = Resource::decl_key(res, face);
+                    face.primitives.send_declare(RoutingContext::with_expr(
+                        Declare {
+                            interest_id,
+                            ext_qos: ext::QoSType::DECLARE,
+                            ext_tstamp: None,
+                            ext_nodeid: ext::NodeIdType::DEFAULT,
+                            body: DeclareBody::DeclareSubscriber(DeclareSubscriber {
+                                id,
+                                wire_expr,
+                                ext_info: sub_info,
+                            }),
+                        },
+                        res.expr(),
+                    ));
+                }
+            } else {
+                for src_face in tables
+                    .faces
+                    .values()
+                    .cloned()
+                    .collect::<Vec<Arc<FaceState>>>()
+                {
+                    if src_face.id != face.id {
+                        for sub in face_hat!(src_face).remote_subs.values() {
+                            if sub.context.is_some() && sub.matches(res) {
+                                let id = if mode.future() {
+                                    let id = face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
+                                    face_hat_mut!(face).local_subs.insert(sub.clone(), id);
+                                    id
+                                } else {
+                                    0
+                                };
+                                let wire_expr = Resource::decl_key(sub, face);
+                                face.primitives.send_declare(RoutingContext::with_expr(
+                                    Declare {
+                                        interest_id,
+                                        ext_qos: ext::QoSType::DECLARE,
+                                        ext_tstamp: None,
+                                        ext_nodeid: ext::NodeIdType::DEFAULT,
+                                        body: DeclareBody::DeclareSubscriber(DeclareSubscriber {
+                                            id,
+                                            wire_expr,
+                                            ext_info: sub_info,
+                                        }),
+                                    },
+                                    sub.expr(),
+                                ));
+                            }
+                        }
+                    }
+                }
+            }
+        } else {
+            for src_face in tables
+                .faces
+                .values()
+                .cloned()
+                .collect::<Vec<Arc<FaceState>>>()
+            {
+                if src_face.id != face.id {
+                    for sub in face_hat!(src_face).remote_subs.values() {
                         let id = if mode.future() {
                             let id = face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
-                            face_hat_mut!(face).local_subs.insert((*res).clone(), id);
+                            face_hat_mut!(face).local_subs.insert(sub.clone(), id);
                             id
                         } else {
                             0
                         };
-                        let wire_expr = Resource::decl_key(res, face);
+                        let wire_expr = Resource::decl_key(sub, face);
                         face.primitives.send_declare(RoutingContext::with_expr(
                             Declare {
                                 interest_id,
@@ -436,172 +477,16 @@ impl HatPubSubTrait for HatCode {
                                     ext_info: sub_info,
                                 }),
                             },
-                            res.expr(),
+                            sub.expr(),
                         ));
                     }
-                } else {
-                    for src_face in tables
-                        .faces
-                        .values()
-                        .cloned()
-                        .collect::<Vec<Arc<FaceState>>>()
-                    {
-                        if src_face.id != face.id {
-                            for sub in face_hat!(src_face).remote_subs.values() {
-                                if sub.context.is_some() && sub.matches(res) {
-                                    let id = if mode.future() {
-                                        let id =
-                                            face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
-                                        face_hat_mut!(face).local_subs.insert(sub.clone(), id);
-                                        id
-                                    } else {
-                                        0
-                                    };
-                                    let wire_expr = Resource::decl_key(sub, face);
-                                    face.primitives.send_declare(RoutingContext::with_expr(
-                                        Declare {
-                                            interest_id,
-                                            ext_qos: ext::QoSType::DECLARE,
-                                            ext_tstamp: None,
-                                            ext_nodeid: ext::NodeIdType::DEFAULT,
-                                            body: DeclareBody::DeclareSubscriber(
-                                                DeclareSubscriber {
-                                                    id,
-                                                    wire_expr,
-                                                    ext_info: sub_info,
-                                                },
-                                            ),
-                                        },
-                                        sub.expr(),
-                                    ));
-                                }
-                            }
-                        }
-                    }
-                }
-            } else {
-                for src_face in tables
-                    .faces
-                    .values()
-                    .cloned()
-                    .collect::<Vec<Arc<FaceState>>>()
-                {
-                    if src_face.id != face.id {
-                        for sub in face_hat!(src_face).remote_subs.values() {
-                            let id = if mode.future() {
-                                let id = face_hat!(face).next_id.fetch_add(1, Ordering::SeqCst);
-                                face_hat_mut!(face).local_subs.insert(sub.clone(), id);
-                                id
-                            } else {
-                                0
-                            };
-                            let wire_expr = Resource::decl_key(sub, face);
-                            face.primitives.send_declare(RoutingContext::with_expr(
-                                Declare {
-                                    interest_id,
-                                    ext_qos: ext::QoSType::DECLARE,
-                                    ext_tstamp: None,
-                                    ext_nodeid: ext::NodeIdType::DEFAULT,
-                                    body: DeclareBody::DeclareSubscriber(DeclareSubscriber {
-                                        id,
-                                        wire_expr,
-                                        ext_info: sub_info,
-                                    }),
-                                },
-                                sub.expr(),
-                            ));
-                        }
-                    }
-                }
-            }
-        }
-        face_hat_mut!(face)
-            .remote_sub_interests
-            .insert(id, (res.as_ref().map(|res| (*res).clone()), aggregate));
-        for dst_face in tables
-            .faces
-            .values_mut()
-            .filter(|f| f.whatami == WhatAmI::Router)
-        {
-            let id = face_hat!(dst_face).next_id.fetch_add(1, Ordering::SeqCst);
-            let options = InterestOptions::KEYEXPRS + InterestOptions::SUBSCRIBERS;
-            get_mut_unchecked(dst_face).local_interests.insert(
-                id,
-                InterestState {
-                    options,
-                    res: res.as_ref().map(|res| (*res).clone()),
-                    finalized: mode == InterestMode::Future,
-                },
-            );
-            let wire_expr = res.as_ref().map(|res| Resource::decl_key(res, dst_face));
-            dst_face.primitives.send_interest(RoutingContext::with_expr(
-                Interest {
-                    id,
-                    mode,
-                    options,
-                    wire_expr,
-                    ext_qos: ext::QoSType::DECLARE,
-                    ext_tstamp: None,
-                    ext_nodeid: ext::NodeIdType::DEFAULT,
-                },
-                res.as_ref().map(|res| res.expr()).unwrap_or_default(),
-            ));
-        }
-    }
-
-    fn undeclare_sub_interest(
-        &self,
-        tables: &mut Tables,
-        face: &mut Arc<FaceState>,
-        id: InterestId,
-    ) {
-        if let Some(interest) = face_hat_mut!(face).remote_sub_interests.remove(&id) {
-            if !tables.faces.values().any(|f| {
-                f.whatami == WhatAmI::Client
-                    && face_hat!(f)
-                        .remote_sub_interests
-                        .values()
-                        .any(|i| *i == interest)
-            }) {
-                for dst_face in tables
-                    .faces
-                    .values_mut()
-                    .filter(|f| f.whatami == WhatAmI::Router)
-                {
-                    for id in dst_face
-                        .local_interests
-                        .keys()
-                        .cloned()
-                        .collect::<Vec<InterestId>>()
-                    {
-                        let local_interest = dst_face.local_interests.get(&id).unwrap();
-                        if local_interest.options.subscribers()
-                            && (local_interest.res == interest.0)
-                        {
-                            dst_face.primitives.send_interest(RoutingContext::with_expr(
-                                Interest {
-                                    id,
-                                    mode: InterestMode::Final,
-                                    options: InterestOptions::empty(),
-                                    wire_expr: None,
-                                    ext_qos: ext::QoSType::DECLARE,
-                                    ext_tstamp: None,
-                                    ext_nodeid: ext::NodeIdType::DEFAULT,
-                                },
-                                local_interest
-                                    .res
-                                    .as_ref()
-                                    .map(|res| res.expr())
-                                    .unwrap_or_default(),
-                            ));
-                            get_mut_unchecked(dst_face).local_interests.remove(&id);
-                        }
-                    }
                 }
             }
         }
     }
+}
 
+impl HatPubSubTrait for HatCode {
     fn declare_subscription(
         &self,
         tables: &mut Tables,
